New changes

+++ app/__init__.py
@@ -106,16 +106,18 @@ def show_one_thing(id):
 def add_a_thing():
     # Get the data from the form
     name  = request.form.get("name")
-    price = request.form.get("price")
 
     # Sanitise the inputs
     name = html.escape(name)
     price = html.escape(price)
 
+    # Get the user id from the session
+    user_id = session["user_id"]
+
     with connect_db() as client:
         # Add the thing to the DB
-        sql = "INSERT INTO things (name, price) VALUES (?, ?)"
-        values = [name, price]
+        sql = "INSERT INTO things (name, user_id) VALUES (?, ?)"
+        values = [name, user_id]
         client.execute(sql, values)
 
         # Go back to the home page
@@ -195,13 +197,28 @@ def login_user():
 @app.get("/delete/<int:id>")
 def delete_a_thing(id):
     with connect_db() as client:
+        # Get our user id from the session
+        user_id = session["user_id"]
+
         # Delete the thing from the DB
-        sql = "DELETE FROM things WHERE id=?"
-        values = [id]
+        # Checking that we are the owner!
+        sql = "DELETE FROM things WHERE id=? AND user_id=?"
+        values = [id, user_id]
         client.execute(sql, values)
 
         # Go back to the home page
-        flash("Thing deleted", "warning")
+        flash("Thing deleted", "success")
         return redirect("/things")
 
 
+#-----------------------------------------------------------
+# Route for user logout
+#-----------------------------------------------------------
+@app.get("/logout")
+def logout():
+    # Clear the session values
+    session.pop("user_id")
+    session.pop("user_name")
+    # Back to the home page
+    flash("You have been logged out", "success")
+    return redirect("/")
